requirements.txt
requests
pandas
pandas_ta
python-dotenv
